@@ -8,17 +8,18 @@
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
+  "type": "module",
   "engines": {
-    "node": "12"
+    "node": "14"
   },
   "dependencies": {
-    "firebase-admin": "^9.5.0",
-    "firebase-functions": "^3.13.2",
-    "node-fetch": "^2.6.1",
+    "firebase-admin": "^9.12.0",
+    "firebase-functions": "^3.15.7",
+    "node-fetch": "^3.0.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "firebase-functions-test": "^0.2.3"
+    "firebase-functions-test": "^0.3.2"
   },
   "private": true
 }
