@@ -34,77 +35,77 @@
     }
   },
   "devDependencies": {
-    "@babel/core": "^7.13.10",
+    "@babel/core": "^7.15.5",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.13.10",
+    "@babel/preset-env": "^7.15.6",
     "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.2.0",
-    "@rollup/plugin-replace": "^2.4.1",
-    "@wdio/cli": "^7.1.2",
-    "@wdio/local-runner": "^7.1.2",
-    "@wdio/mocha-framework": "^7.1.2",
-    "@wdio/sauce-service": "^7.1.1",
-    "@wdio/selenium-standalone-service": "^7.1.1",
-    "@wdio/spec-reporter": "^7.1.1",
+    "@rollup/plugin-commonjs": "^20.0.0",
+    "@rollup/plugin-node-resolve": "^13.0.5",
+    "@rollup/plugin-replace": "^3.0.0",
+    "@wdio/cli": "^7.13.2",
+    "@wdio/local-runner": "^7.13.2",
+    "@wdio/mocha-framework": "^7.13.2",
+    "@wdio/sauce-service": "^7.13.2",
+    "@wdio/selenium-standalone-service": "^7.13.2",
+    "@wdio/spec-reporter": "^7.13.2",
     "babel-eslint": "^10.1.0",
     "body-parser": "^1.19.0",
-    "chalk": "^4.1.0",
-    "chromedriver": "^89.0.0",
-    "core-js": "^3.9.1",
-    "cssnano": "4.1.10",
+    "chalk": "^4.1.2",
+    "chromedriver": "^94.0.0",
+    "core-js": "^3.18.1",
+    "cssnano": "5.0.8",
     "dom-utils": "^0.9.0",
     "dynamic-import-polyfill": "^0.1.1",
-    "eslint": "^7.21.0",
+    "eslint": "^7.32.0",
     "eslint-config-google": "^0.14.0",
     "express": "^4.17.1",
-    "firebase-tools": "^9.6.1",
-    "fs-extra": "^9.1.0",
-    "globby": "^11.0.2",
+    "firebase-tools": "^9.19.0",
+    "fs-extra": "^10.0.0",
+    "globby": "^12.0.2",
     "gm": "^1.23.1",
     "gulp": "^4.0.2",
     "gzip-size": "^6.0.0",
     "he": "~1.2.0",
-    "highlight.js": "^10.6.0",
+    "highlight.js": "^11.2.0",
     "html-minifier": "^4.0.0",
-    "husky": "^5.1.3",
+    "husky": "^7.0.2",
     "idlize": "^0.1.1",
-    "imagemin": "^7.0.1",
+    "imagemin": "^8.0.1",
     "imagemin-pngquant": "^9.0.2",
     "jsesc": "^3.0.2",
-    "markdown-it": "^12.0.4",
-    "markdown-it-anchor": "^7.1.0",
+    "markdown-it": "^12.2.0",
+    "markdown-it-anchor": "^8.3.1",
     "md5": "^2.3.0",
-    "mocha": "^8.3.1",
+    "mocha": "^9.1.2",
     "moment-timezone": "^0.5.33",
     "morgan": "^1.10.0",
-    "node-fetch": "^2.6.1",
+    "node-fetch": "^3.0.0",
     "nunjucks": "^3.2.3",
     "path": "^0.12.7",
-    "postcss": "^8.2.8",
-    "postcss-import": "^14.0.0",
+    "postcss": "^8.3.8",
+    "postcss-import": "^14.0.2",
     "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "regenerator-runtime": "^0.13.7",
+    "postcss-url": "^10.1.3",
+    "regenerator-runtime": "^0.13.9",
     "require-dir": "^1.2.0",
     "rev-hash": "^3.0.0",
     "rev-path": "^2.0.0",
-    "rollup": "^2.41.1",
+    "rollup": "^2.57.0",
     "rollup-plugin-terser": "^7.0.2",
     "superstatic": "^7.1.0",
-    "wdio-chromedriver-service": "^7.0.0",
-    "web-vitals": "^1.1.1",
-    "webdriverio": "^7.1.1",
+    "wdio-chromedriver-service": "^7.2.2",
+    "web-vitals": "^2.1.0",
+    "webdriverio": "^7.13.2",
     "whatwg-fetch": "^3.6.2",
-    "workbox-background-sync": "^6.1.2",
-    "workbox-broadcast-update": "^6.1.2",
-    "workbox-core": "^6.1.2",
-    "workbox-expiration": "^6.1.2",
-    "workbox-google-analytics": "^6.1.2",
-    "workbox-precaching": "^6.1.2",
-    "workbox-routing": "^6.1.2",
-    "workbox-strategies": "^6.1.2",
-    "workbox-streams": "^6.1.2",
-    "workbox-window": "^6.1.2"
+    "workbox-background-sync": "^6.3.0",
+    "workbox-broadcast-update": "^6.3.0",
+    "workbox-core": "^6.3.0",
+    "workbox-expiration": "^6.3.0",
+    "workbox-google-analytics": "^6.3.0",
+    "workbox-precaching": "^6.3.0",
+    "workbox-routing": "^6.3.0",
+    "workbox-strategies": "^6.3.0",
+    "workbox-streams": "^6.3.0",
+    "workbox-window": "^6.3.0"
   }
 }
