@@ -1,13 +1,17 @@
-const {log: v1} = require('./v1/log');
-const {log: v2} = require('./v2/log');
+import {v1} from './log/v1.js';
+import {v2} from './log/v2.js';
+import {v3} from './log/v3.js';
 
 /**
  * @param {Object} request
  * @param {Object} response
  */
-async function log(request, response) {
+export async function log(request, response) {
   const version = Number(request.query.v);
   switch (version) {
+    case 3:
+      await v3(request);
+      break;
     case 2:
       await v2(request);
       break;
@@ -17,7 +21,3 @@ async function log(request, response) {
   }
   response.end();
 }
-
-module.exports = {
-  log,
-};
