@@ -1,9 +1,10 @@
-const fs = require('fs-extra');
-const gulp = require('gulp');
-const {resetManifest} = require('./utils/assets');
-const config = require('../config');
+import fs from 'fs-extra';
+import gulp from 'gulp';
+import {resetManifest} from './utils/assets.js';
 
 gulp.task('clean', async () => {
+  const config = await fs.readJSON('./config.json');
+
   await fs.remove(config.publicDir);
   resetManifest();
 });
