@@ -1,9 +1,9 @@
-const gulp = require('gulp');
-const server = require('./utils/server');
+import gulp from 'gulp';
+import {start} from './utils/server.js';
 
 gulp.task('server', async () => {
   try {
-    await server.start();
+    await start();
   } catch (err) {
     console.error(err);
   }
