@@ -1,3 +1,1 @@
-const ENV = process.env.NODE_ENV || 'development';
-
-module.exports = {ENV};
+export const ENV = process.env.NODE_ENV || 'development';
