@@ -1,1 +1,9 @@
-require('require-dir')('./tasks');
+import './tasks/build.js';
+import './tasks/clean.js';
+import './tasks/content.js';
+import './tasks/css.js';
+import './tasks/images.js';
+import './tasks/javascript.js';
+import './tasks/server.js';
+import './tasks/sw.js';
+import './tasks/watch.js';
